+++ src/utils/Slice.js
@@ -1,5 +1,5 @@
-import {createSlice} from '@reduxjs/toolkit';
-import { fetchAuthMappedMenuByThunk } from './thunk';
+import {createAsyncThunk, createSlice} from '@reduxjs/toolkit';
+import { addAuthApi, getAuthGroupApi } from '../api/authgroup';
 //jane
 const companyMgmtInitialState = {
     companyInfo: {
@@ -138,45 +138,48 @@ export const modalSlice = createSlice({
       }
     });
   }
-
-/** @author subo */
-export const authMappedMenuSlice = createSlice({
-  name: 'authMappedMenu',
-  initialState: {
-    data: [],
-    isLoading: false,
-    error: null,
-  },
-  reducers: {
-    startLoading: (state) => {
-      state.isLoading = true;
+  export const fetchAuthGroups = createAsyncThunk(
+    'authGroup/fetchAuthGroups',
+    async (queryParams) => {
+      try {
+        const response = await getAuthGroupApi({ params: queryParams });
+        return response.data;
+      } catch (error) {
+        throw Error(error);
+      }
+    }
+  );
+  export const addAuthGroup = createAsyncThunk('authGroup/addAuthGroup', async (newAuthGroup, { dispatch }) => {
+    await addAuthApi({ data: newAuthGroup });
+    dispatch(fetchAuthGroups()); // 새 권한을 추가한 후 목록을 새로고침
+  });
+  export const authGroupSlice = createSlice({
+    name: 'authGroup',
+    initialState: {
+      data: [],
+      hasMore: true,
+      isLoading: false,
+      error: null,
     },
-    hasError: (state, action) => {
-      state.isLoading = false;
-      state.error = action.payload;
-    },
-    authMappedMenuReceived: (state, action) => {
-      state.isLoading = false;
-      state.data = action.payload;
-    },
-  },  
-  extraReducers: (builder) => {
-    builder
-      .addCase(fetchAuthMappedMenuByThunk.pending, (state) => {
+    reducers: {
+      fetchDataStart: (state) => {
         state.isLoading = true;
-      })
-      .addCase(fetchAuthMappedMenuByThunk.fulfilled, (state, action) => {
+      },
+      fetchDataSuccess: (state, action) => {
         state.isLoading = false;
-        state.data = action.payload;
-      })
-      .addCase(fetchAuthMappedMenuByThunk.rejected, (state, action) => {
+        state.data = [...state.data, ...action.payload.data];
+        state.hasMore = action.payload.hasMore;
+      },
+      fetchDataFailure: (state, action) => {
         state.isLoading = false;
-        state.error = action.error.message;
-      });
-  },
-});
-
-
+        state.error = action.payload;
+      },
+      resetData: (state) => {
+        state.data = [];
+        state.hasMore = true;
+      },
+    },
+  });
 
 export const {menu, favor, recent, profileList, empId, compId} = menuSlice.actions;
 export const {load} = recentSlice.actions;
